Change map structure
Change in to the map structure.
Spiltup the actual workspace and the  documentation part.
This makes it possible to offer a seperate zip of the workspace to the end user.
The workspace has the same name in both the SLI and the local programs.
This is possible because they are sharing the same TxtDeps and toolchainin the projects